+++ src/editor/core/event/handlers/drop.ts
@@ -1,13 +1,16 @@
+import { isPromiseFunction } from '../../../utils'
+import { IOverrideResult } from '../../override/Override'
 import { CanvasEvent } from '../CanvasEvent'
 import { pasteImage } from './paste'
 
-export function drop(evt: DragEvent, host: CanvasEvent) {
+export async function drop(evt: DragEvent, host: CanvasEvent) {
   const draw = host.getDraw()
   // 自定义拖放事件
   const { drop } = draw.getOverride()
   if (drop) {
-    drop(evt)
-    return
+    const overrideResult = isPromiseFunction(drop) ? await drop(evt) : drop(evt)
+    // 默认阻止默认事件
+    if ((<IOverrideResult>overrideResult)?.preventDefault !== false) return
   }
   evt.preventDefault()
   const data = evt.dataTransfer?.getData('text')