+++ src/editor/core/override/Override.ts
@@ -1,5 +1,19 @@
+export interface IOverrideResult {
+  preventDefault?: boolean
+}
+
 export class Override {
-  public paste: ((evt?: ClipboardEvent) => void) | undefined
-  public copy: (() => void) | undefined
-  public drop: ((evt: DragEvent) => void) | undefined
+  public paste:
+    | ((
+        evt?: ClipboardEvent
+      ) => void | Promise<void> | IOverrideResult | Promise<IOverrideResult>)
+    | undefined
+  public copy:
+    | (() => void | Promise<void> | IOverrideResult | Promise<IOverrideResult>)
+    | undefined
+  public drop:
+    | ((
+        evt: DragEvent
+      ) => void | Promise<void> | IOverrideResult | Promise<IOverrideResult>)
+    | undefined
 }